+++ src/main/java/web/dao/UserDao.java
@@ -8,8 +8,8 @@ public interface UserDao {
 
     void add(User user);
     List<User> show();
-    User getUser(int id);
-    void update(int id);
-    void delete(int id);
+    User getUser(Long id);
+    void update(Long id);
+    void delete(Long id);
 
 }
