+++ src/main/java/web/service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserService {
     void add(User user);
     List<User> show();
-    User getUser(int id);
-    void update(int id);
-    void delete(int id);
+    User getUser(Long id);
+    void update(Long id);
+    void delete(Long id);
 }
